requirement.txt
streamlit==1.39.0
langchain==0.1.10
langchain-openai==0.0.8
streamlit
google-generativeai
python-dotenv
pdf2image